backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
sqlalchemy
psycopg2-binary
authlib
srt
webvtt-py
passlib
bcrypt
azure-cognitiveservices-speech
azure-storage-blob
pandas

# Database dependencies
asyncpg    
alembic
